more work on ground_script
