+++ ld_script.txt
@@ -347,7 +347,6 @@ SECTIONS {
         src/code_809C5C4.o(.text);
         asm/code_809C5C4.o(.text);
         src/code_809D148.o(.text);
-        asm/code_809D148.o(.text);
         src/ground_script.o(.text);
         asm/handleaction.o(.text);
         src/ground_script_1.o(.text);
